RecOrigen/AccountCreator/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
fake-useragent==1.5.1
frozenlist==1.4.1
h11==0.14.0
idna==3.9
multidict==6.1.0
outcome==1.3.0.post0
PGPy==0.6.0
protonmail-api-client==1.7.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.20.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.24.0
setuptools==74.1.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
yarl==1.11.1
